WIP image focus/wp-view + module refactoring
@TODO : absolutely find a better than focus/wp-list/thing

doesn't support changing the focus yet. Lots of stuff to cleanup and
rewrite, this is still very exploratory.

gets rid of Index.elm... I think it doesn't really make sens and was
really annoying. Instead all the views are in the `Main.elm` file. I
made a separate `Model.elm` for clarity and because I might want to
refactor stuff depending on the `Msg` type in the future.